it should be a float already

+++ init/radiosonde_ACCACIAconversions.py
@@ -712,11 +712,11 @@ def loadAircraft(data):
     mi_number = data['2DS'].variables['NC_MI'][:]
     psd_number = data['2DS'].variables['PSD_Num_E'][:]
 
-    hi_number = hi_number.astype(np.float)
-    mi_number = mi_number.astype(np.float)
+    # hi_number = hi_number.astype(np.float)
+    # mi_number = mi_number.astype(np.float)
 
-    hi_number[hi_number==-9999.] = np.nan
-    mi_number[mi_number==-9999.] = np.nan
+    hi_number[hi_number==-9999] = np.nan
+    mi_number[mi_number==-9999] = np.nan
 
     data['Aircraft']['Nice'] = hi_number + mi_number + np.nansum(psd_number[:,9:],1)
